+++ lib/app/customers/workout/[workout]_page.dart
@@ -1,11 +1,11 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:rivinha_fitness/app/config/app_provider.dart';
+import 'package:rivinha_fitness/modules/core/config/app_provider.dart';
 import 'package:rivinha_fitness/app/customers/workout/workout_store.dart';
 import 'package:rivinha_fitness/app/widgets/list_item.dart';
-import 'package:rivinha_fitness/model/workout.dart';
-import 'package:rivinha_fitness/routes.dart';
+import 'package:rivinha_fitness/modules/core/config/routes.dart';
+import 'package:rivinha_fitness/modules/workout/data/models/workout_model.dart';
 import 'package:routefly/routefly.dart';
 
 class CustomerWorkoutPage extends StatefulWidget {
@@ -26,7 +26,7 @@ class _CustomerWorkoutPageState extends State<CustomerWorkoutPage> {
     var workoutExists = paramValue != null;
 
     if (workoutExists) {
-      Workout workout = Workout.fromJson(jsonDecoder.convert(paramValue));
+      WorkoutModel workout = WorkoutModel.fromJson(jsonDecoder.convert(paramValue));
       store.selectWorkout(workout: workout);
     }
 
@@ -55,8 +55,7 @@ class _CustomerWorkoutPageState extends State<CustomerWorkoutPage> {
               children: [
                 Padding(
                   padding: const EdgeInsets.only(bottom: 20),
-                  child: Text(
-                      store.workoutSelected!.description ?? 'Sem descrição',
+                  child: Text(store.workoutSelected!.description ?? 'Sem descrição',
                       style: const TextStyle(
                         fontSize: 18,
                       )),
@@ -65,12 +64,9 @@ class _CustomerWorkoutPageState extends State<CustomerWorkoutPage> {
                       padding: const EdgeInsets.symmetric(vertical: 2),
                       child: ListItem(
                         title: e.name,
-                        description:
-                            'Séries: ${e.sets}   Repetições: ${e.reps}   Peso: ${e.weight ?? '-'}',
+                        description: 'Séries: ${e.sets}   Repetições: ${e.reps}   Peso: ${e.weight ?? '-'}',
                         onTap: () {
-                          Routefly.push(routePaths
-                              .customers.workout.exercise.$item
-                              .changes({
+                          Routefly.push(routePaths.customers.workout.exercise.$item.changes({
                             'item': jsonEncoder.convert(e),
                           }));
                         },
