+++ lib/app/customers/workout/exercise/new_exercise_store.dart
@@ -3,10 +3,10 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:mobx/mobx.dart';
-import 'package:rivinha_fitness/app/config/app_provider.dart';
+import 'package:rivinha_fitness/modules/core/config/app_provider.dart';
 import 'package:rivinha_fitness/app/customers/workout/workout_store.dart';
-import 'package:rivinha_fitness/model/exercise.dart';
-import 'package:rivinha_fitness/routes.dart';
+import 'package:rivinha_fitness/modules/core/config/routes.dart';
+import 'package:rivinha_fitness/modules/workout/data/models/exercise_model.dart';
 import 'package:routefly/routefly.dart';
 import 'package:uuid/uuid.dart';
 part 'new_exercise_store.g.dart';
@@ -26,8 +26,7 @@ abstract class _NewExerciseStoreBase with Store {
 
   @action
   getImage() async {
-    final pickedFile =
-        await ImagePicker().pickImage(source: ImageSource.gallery);
+    final pickedFile = await ImagePicker().pickImage(source: ImageSource.gallery);
 
     if (pickedFile != null) {
       image = File(pickedFile.path);
