+++ lib/app/customers/workout/workout_store.dart
@@ -1,29 +1,28 @@
 import 'package:mobx/mobx.dart';
-import 'package:rivinha_fitness/model/exercise.dart';
-import 'package:rivinha_fitness/model/workout.dart';
+import 'package:rivinha_fitness/modules/workout/data/models/exercise_model.dart';
+import 'package:rivinha_fitness/modules/workout/data/models/workout_model.dart';
 part 'workout_store.g.dart';
 
 class WorkoutStore = _WorkoutStoreBase with _$WorkoutStore;
 
 abstract class _WorkoutStoreBase with Store {
   @observable
-  Workout? workoutSelected;
+  WorkoutModel? workoutSelected;
 
   @action
-  void selectWorkout({required Workout workout}) {
+  void selectWorkout({required WorkoutModel workout}) {
     workoutSelected = workout;
   }
 
   @action
   addExerciseToWorkout({required ExerciseModel exercise}) {
-    var exerciseExists = workoutSelected!.exercises
-        .firstWhere((element) => element.id == exercise.id);
+    var exerciseExists = workoutSelected!.exercises.firstWhere((element) => element.id == exercise.id);
 
     if (exerciseExists.id == null) {
       workoutSelected!.exercises.add(exercise);
     } else {
-      workoutSelected!.exercises[workoutSelected!.exercises
-          .indexWhere((element) => element.id == exercise.id)] = exercise;
+      workoutSelected!.exercises[workoutSelected!.exercises.indexWhere((element) => element.id == exercise.id)] =
+          exercise;
     }
   }
 }
