+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:rivinha_fitness/app/config/app.dart';
-import 'package:rivinha_fitness/app/config/app_provider.dart';
+import 'package:rivinha_fitness/modules/core/config/app.dart';
+import 'package:rivinha_fitness/modules/core/config/app_provider.dart';
 
 Future<void> main() async {
   await setupApp();
