+++ lib/modules/core/vos/value_object.dart
