+++ lib/modules/workout/domain/entities/workout.dart
@@ -1,0 +1,15 @@
+import 'package:rivinha_fitness/modules/workout/data/models/exercise_model.dart';
+
+class Workout {
+  final String id;
+  final String name;
+  final String? description;
+  final List<ExerciseModel> exercises;
+
+  Workout({
+    required this.id,
+    required this.name,
+    this.description,
+    required this.exercises,
+  });
+}
