+++ pubspec.yaml
@@ -57,6 +57,7 @@ dependencies:
   firebase_crashlytics: ^3.4.9
   cloud_firestore: ^4.14.0
   firebase_storage: ^11.6.0
+  dartz: ^0.10.1
 
 dev_dependencies:
   flutter_test:
